Add support for for different modes of display in formatNumber
This adds a new `display` option to the formatNumber function:
- 0 = use short suffixes (default, e.g. 61.492 Oc)
- 1 = use full name (e.g., "61.492 Octillion")
- 2 = use scientific notation (e.g., "6.149e28")

This commit isn't ready to be pulled immediately, there should be the /settings command to configure this. It would work, because it defaults to 0, so it wouldn't change anything, so you can if you want, but ideally the settings command would be done so this is actually somewhat useful.

if you want, you can change some uses of the formatnumber function to use the new method so the formatting is nicer? like, as an example, you could change leaderboard to use the full name instead of the shortened version. Just an example, not saying you should.